Update index.html
Update Beta 0.0.2: With this update, I've kept the same features, but reduced RAM usage and loading speeds! Essentially, I have removed the floating widget. This widget does not make sense to have, as it is very easy to scroll to the top or the bottom of the screen to book a conversation.